camera.jl

camera.jl

Camera script with defines camera object, its functions, events and creates a single camera object for whole szene.

App.rezizeWindow — Method.

sets glfw window size + viewport

source

App.Camera — Type.

camera object with holds position, rotation, scaling and various matrices like MVP

source

App.forward — Method.

gets forward vector of camera direction

source

App.right — Method.

gets right vector of camera direction

source

App.up — Method.

gets up vector of camera direction

source

App.setProjection — Method.

sets projection matrix

source

App.setView — Method.

sets view matrix

source

App.OnKey — Method.

event which catches keyboard inputs. here keys for wireframe, fullscreen and camera movement are defined

source

App.OnMouseKey — Method.

event which catches mouse key inpits and hides/shows cursor when mouse button is pressed

source

App.OnCursorPos — Method.

event which catches mouse position for camera rotation event

source

App.rotate — Method.

rotates camera

source

App.move — Method.

moves camera, adds vector to current position

source

App.OnRotate — Method.

event which calculates cursor position shifts and calls rotate function

source

App.setPosition — Method.

sets camera position

source

App.OnMove — Method.

event which updates positions shifts (left,right,up,down,forward,back) key is (left,right,up,down,forward,back) m is direction value with weight (positive, negative)

source

App.Update — Method.

update function where camera translation is update only when camera was moved by input. here cameras MVP Matrix is updated aswell

source

App.OnUpdate — Method.

event which is called by game loop and calls real update function this event resets camera moved state

source